Updated to 2014/08/26 development build * Added read-only YU-RIS YCG image format support (requested by ZS) * Added full YU-RIS v4.90 archive sub-format support and fixed bug in v3.00 creation algorithm (requested by ZS) * Added experimental read-only Nitro+ NPA archive format support (Chaos;Head only, ported from NIPA sources by Nik)

+++ src/AnimED_Progress.pas
@@ -18,8 +18,8 @@ const pcolDefault = $00D08410;
       pcolWrite   = $00C080FF; //to-do: find it =_=
                                //update: FOUND ! ^__^
 
-procedure Progress_Max(i : integer; ColorMode : TColor = pcolDefault);
-procedure Progress_Pos(i : integer);
+procedure Progress_Max(i : int64; ColorMode : TColor = pcolDefault);
+procedure Progress_Pos(i : int64);
 //procedure Progress_Col(ColorMode : TColor);
 
 implementation
